Merging the sandbox RFC2 into the trunk http://svn.osgeo.org/metacrs/csmap/sandbox/RFC2/
This will have a short term impact on the coordinate system transformations involving grid file transformations, meaning that such transformations will not work during the rest of the implementation of RFC2.
The transformations involved are ATS77, RGF93 and Japan.

Everything would get back to normal by the end of September.