CS-MAP Trac # 88 -- Changes which produced revision 1590 in the trunk are submitted to this branch.
This submission enhances the ability of CS-MAP with regard to whole world maps, especially with regard to the Pseudo-Mercator projection which is a projection variation which is specifically designed for use with on-line data sources such as Google, Bing, etc.

In the Mercator projection, either pole is singularity point; i.e. a point which cannot exist on the map.  (If we tried, a pole would essentially be a horizontal line at infinity.)  However, most users expect to use the Pseudo-Mercator variation of CS-MAP's implementation of the Mercator projection to view whole world maps and images generated by Google, Bing, and other on-line map services. 

There were two problems encountered in attempting to accommodate these users.

1> In the Mercator (all three variations), the Eckert IV, the Eckert VI, the Miller, and the American Polyconic projections, the CS-MAP implementation would not consider the anti-meridian (central meridian plus/minus 180 degrees) as being with the domain of the projection.  As a result, a map or image whose longitude extents include +/- 180 would cause an exception.  This was a coding error and was corrected quite easily.  The both anti-meridian longitude values  are now considered to part of the projection domain, and also the default useful range of any CRS based on these projections.

2> In addition to frequently containing points and/or bounding boxes referencing the anti-meridian, the on-line data sources often reference the poles in their data and/or bounding boxes.  As described above, this would cause an exception whenever the Pseudo-Mercator projection would be used.  While technically incorrect, but in order to accommodate the users of on-line data sources, the Pseudo variation of CS-MAP implementation of the Mercator projection has been changed to:
a> consider either pole as within the domain of the Mercator projection,
b> limit the northing return value (prior to false origin application) to plus/minus 20,000,000 meters (about 85 degrees of latitude), and
c> replace all latitudes in the range of greater than 85 but less than or equal to 90 to be 85 degrees.

In this manner, the Pseudo Mercator projection can be used to view most all whole world maps without producing an exception.