Trac Tickets: 91, 94, 95, and a lot of help for #45.
The changes submitted with this maintenance submission are made to address two general problems for which a defect does not exist.

1> WKT produced by Spatial Resources were not being processed correctly.  While it appeared that releases prior to RFC 2 were processing Spatial Resource WKT properly, this is not precisely true.  Only WKT referenced to a small number of datum were being processed correctly; the remainder were being processed erroneously.  In any case, these changes correct both situations and the post RFC 2 CS-MAP (i.e. 13.0) will process all WKT produced by Spatial Resources correctly or not at all.

2> Our testing programs have been producing a fair number of bogus error reports  making it difficult to actually know which, if any, are legitimate error reports.  Thus, a substantial effort has been made to eliminate all bogus errors, and having accomplished much in doing so, correct the real errors thus identified.

The following files are compiled to produce the NameMapper.csv file.
	Data/EllipsoidKeyNameMap.csv
	Data/ProjectiveKeyNameMap.csv

ELiminating error reports from the COnsoleTestCpp facility required the 
correction of one ellipsoid, one datum, and 11 coordinate systems.
	Dictionaries/category.asc
	Dictionaries/coordsys.asc
	Dictionaries/datums.asc
	Dictionaries/elipsoid.asc
	Dictionaries/NameMapper.csv

The results produced by the CS_defCmp function were dependent upon the order of 
the parameters provided which was considered to be a defect: 
	Source/CS_defCmp.c

The ability to consider various combinations of projections as being equivalent 
was improved in the CS_defCmpEx functions.  Example, MRCAT and MRCATK are simply 
different parameterizations of the same thing.  Improving this ability 
eliminated several bogus error reports from the ConsoleTestCpp Facility.
	Source/CS_defCmpEx.c

WKT strings produced by SpatialReferences.org were not being processed properly. 
 This is true as there is no explicit flavor for this source of WKT and thus the 
names of datums in the WKT could not be napped by the name mapper.  However, 
this particular source of WKT always include EPSG AUTHORITY elements.  So 
CS_wkt2CsEx was improved to use the EPSG codes contained in the WKT to obtain a 
CS-MAP datum name and thus improve the ability of the WKT processor to handle 
WKT produced by sources other those which have been established as a recognized 
flavor.
	Source/cs_wellknowntext.cpp

Many of the false error reports issued by ConsoleTestCpp were due to errors in 
the cs_Csdef_ equivalent the EPSG sub-module produces.  Correcting this errors 
substantially redued the number of bogus error messages produced.
	Source/csEpsgStuff.cpp

Coordinate System Definition Changes:
	Ellipsoid: NWL-9D                replaced by  NWL-9D/1
	    Datum: NWL                   replaced by  NWL/1
          CRS: HARN/HI.HI-1F         replaced by  HARN/HI.HI-1Fa
          CRS: HARN/HI.HI-2F         replaced by  HARN/HI.HI-2Fa
          CRS: HARN/HI.HI-3F         replaced by  HARN/HI.HI-3Fa
          CRS: HARN/HI.HI-4F         replaced by  HARN/HI.HI-4Fa
          CRS: HARN/HI.HI-5F         replaced by  HARN/HI.HI-5Fa
          CRS: Nahrwan67.UTM-37N     replaced by  Nahrwan67.UTM-37N/1
          CRS: REUNION.Reunion/TM/02 replaced by  REUNION.Reunion/TM/03