Fix "Updating user definitions not possible if CS_usrDir is set" as part of RFC6
The central [CS_DefinitionUpdate()] function in csIoUtil.hpp is supposed to check, whether the definition passed in, can be updated in either the system or the user CSD file. It does that by calling [CS_DescribeT()] which checks whether the definition exist and is not write-protected. CS_DescribeT is (supposed) being called for the system file & the user file. However, there was a typo in causing the system CSD file being checked 2 times.

Ultimately, the definition was thus always considered "not 
yet existing" and was always added (again) - no matter whether the actual 
changes was an update of a user definition. The GIS platform asserted on this, actually.
That is, definitions which were supposed to get updated only, were returned
'added' by CSMAP.

The fix was just to get the 2nd check right - see changes in [csIoUtil.hpp].

Secondly, the defect revealed another problem in how categories were not written 
back to disk by the CS_ctdel() but CS_ctupd() does so. The CS_ct*() functions are supposed to behave like all the other CS_[xx]del/upd() functions. Note, that this is not a problem in the GIS 
Platform's MgGeometry.dll because there an additional call is made to make the 
changes persistent. Anyway, I've now changed CSMAP to always update the CSD 
content in case modifications have been made. The GIS Platform I'll modify 
separately.

Reviewed by Hugues.